Extend libpolicy to provide access to all device policy fields
This CL extends libpolicy so that it provides access to all fields of
the device policy protobuf. That also includes the ephemeral users
policy which is being added to the protobuf on the Chrome side in the
following CL:

https://chromiumcodereview.appspot.com/9104038/

This CL should be committed once the above CL goes in and the version
of the chromeos-base/protofiles ebuild has been bumped to pull in the
new protobuf files.

BUG=chromium-os:20004
TEST=libpolicy_unittest
CQ-DEPEND=Ie7e0e5e1ee4deb4e2cce46702e3d865f27efcfce

Change-Id: I09616bbbe5b57815ce377d209444c8c1876e9499
Reviewed-on: https://gerrit.chromium.org/gerrit/15147
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Tested-by: Bartosz Fabianowski <bartfab@google.com>
Commit-Ready: Bartosz Fabianowski <bartfab@google.com>