[cromo] fix SmsMessageFragment::CreateFragment export on 64bit systems
When I built up the list of symbols to export, I looked at `readelf -s`
passed through `c++filt` which doesn't tell the whole story.  In this
case, the signature for a function with a size_t was written as unsigned
int but 64bit has unsigned long, so it didn't get exported.

Ideally we'd use size_t in the linker script, but the linker doesn't
understand that type.  Not sure if bug or by design, but not that big
of a deal for our needs.

BUG=None
TEST=`emerge-amd64-generic cromo && readelf -s` shows exported symbol
TEST=`emerge-x86-generic cromo && readelf -s` shows exported symbol

Change-Id: I8ea07db7b6103dc756a790db152fe73eacceb2c0
Reviewed-on: https://gerrit.chromium.org/gerrit/15437
Reviewed-by: Jason Glasgow <jglasgow@chromium.org>
Reviewed-by: ttuttle <ttuttle@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>