Touch .developer_mode in stateful if the dev switch is enabled.
When installing a non-developer image with the dev mode switch
enabled, touch /mnt/stateful_partition/.developer_mode; this
prevents the subsequent boot flow from triggering a wipe of the
stateful partition for the transition to developer mode.

This case normally can only occur if installing a standard
(non-developer) image via recovery with the dev-mode switch enabled.
The code is being added so that similar code in initramfs can be
deleted.

BUG=chromium-os:22614
TEST=run via recovery with the dev switch enabled

Change-Id: Ieeb8ed9bb68507e44cbf7d30af5085ca258a1ccb
Reviewed-on: https://gerrit.chromium.org/gerrit/15535
Reviewed-by: Paul Taysom <taysom@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>