Fix user feedback items
Some items had wrong paths or not path, and some probably no longer work
at all. This is just cleanup.

Fix some errors in the chrome:system menus

Not all menu items chrome:system worked. Partly it was due to simple
differences, such as referring to /sbin/lsmod when on busybox it is
/bin/lsmod; or assuming that something was in the path (lsusb)
when it was not (the chromium path is /bin:/usr/bin:/usr/bin/X11)

I added an 'env' command to make it easier for users to see just what
was what in the PATH and other variables.

removed whitespace at end of one comment

Added support for powerm files

remove trailing white space

BUG=chrome-os-partner:7911
TEST=boot chromeos and verify that all the chrome:system bits now work.
Change-Id: I78048a147ea3463f23074bb0036ec63cd19bbfca
Reviewed-on: https://gerrit.chromium.org/gerrit/15651
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>
Commit-Ready: Ronald G. Minnich <rminnich@chromium.org>
Tested-by: Ronald G. Minnich <rminnich@chromium.org>