Reduce number of resume attempts and delay between attempts.
This reduces the number of resume attempts from 20 to 3 (total
4 attempts), and the delay between a failure and the subsequent attempt
from 60 to 20 seconds.  In total, the delay induced due to resume
attempts will not exceed 1 minute.  This appears long enough to overcome
transient connectivity issues such as a user moving between access
points, a temporary Ethernet cable disconnect, or a switch restarting.
It is necessary for reducing the amount of time a user needs to wait in
the case of interrupted network connectivity during OOBE update.

This requires a corresponding fix to unittests, as the assumption on the
number of retries has changed.

BUG=chrome-os-partner:7961
TEST=Passed unittests; verifying OOBE behavior pending at test lab.

Change-Id: I669ac6e2a1264c1d154b23a06c66763f155ab946
Reviewed-on: https://gerrit.chromium.org/gerrit/15639
Reviewed-by: Gilad Arnold <garnold@chromium.org>
Tested-by: Gilad Arnold <garnold@chromium.org>
Commit-Ready: Gilad Arnold <garnold@chromium.org>