shill: (wifi) only register endpoints for which we have a service.
When we process a BSSRemoved event, we check that we know which Service the
Endpoint corresponds to. Previously, however, this check could fail.

The failure occurred in cases where the Endpoint was registered, but no
corresponding Service was found (or created). Resolve this problem by
registering the Endpoint only after we have a corresponding Service.

Also, remove some special-case code which is no longer needed.

BUG=chromium-os:26159
TEST=new unit tests

Change-Id: Ib4c13f9e426f69cd979afb1efdd4faad80fd853d
Reviewed-on: https://gerrit.chromium.org/gerrit/15697
Commit-Ready: mukesh agrawal <quiche@chromium.org>
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>