Enable RAR archive support.
This CL registers the RAR file extension in the archive manager in
order to enable the mounting of RAR archives as a virtual filesystem,
which can then be viewed in the Chromium OS File Manager.

The seccomp filter policy files are also updated to allow the system
call 'ioctl(cmd==TCGETS)', which is triggered by the isatty check in
the unrar code.

BUG=chromium-os:26687
TEST=Tested the following:
1. Build cros-disks for {x86,amd64,arm}-generic.
2. Run cros-disks unit tests.
3. Test mounting and unmounting RAR archives using cros-disks DBus API
   on a Cr48.

Change-Id: I30e24355194930bfb9b8c7b214af50d305691f47
Reviewed-on: https://gerrit.chromium.org/gerrit/16418
Reviewed-by: Kees Cook <keescook@chromium.org>
Reviewed-by: Darin Petkov <petkov@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Commit-Ready: Ben Chan <benchan@chromium.org>