[cryptohome] Remove legacy vault support.
The legacy vault format was deprecated on 2010-07-19 by fes; we have never
shipped a machine to users that used this vault format, and it would only have
been present on extremely old dogfood hardware. We don't care any more. Removing
it leads to a bunch of cascading refactors which clean a lot of other code up.

TEST=unit,trybot
BUG=None

Change-Id: I1ba3e95363d02f957be666bed00fd2c390b56a5f
Signed-off-by: Elly Jones <ellyjones@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/16302
Reviewed-by: Kees Cook <keescook@chromium.org>
Reviewed-by: Will Drewry <wad@chromium.org>