[cryptohome] remove fallback_to_scrypt = false
With fallback_to_scrypt = false, we encrypt the vkk with the user's raw
passphrase (not hashed or anything), which provides no brute-force protectiona
at all. We never want this mode, and indeed the only current uses of it are in
unit tests.

Removing this flag causes the encrypt/decrypt methods in the Crypto class to
split into functions that do two disjoint things depending on their encryption
mode, so we can split them out into smaller functions.

TEST=unit
BUG=None

Change-Id: If0a3fcdddb1f652af527369151bd2d6d6eb393c5
Signed-off-by: Elly Jones <ellyjones@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/16360
Reviewed-by: Kees Cook <keescook@chromium.org>
Reviewed-by: Will Drewry <wad@chromium.org>