shill: Add code to wait for dhcpcd to exit before continuing.
In the past, we stop dhcpcd by sending it a SIGTERM without waiting for
it to exit.  This cause a race condition where the first instance of
dhcpcd is still exiting when we start the second instance.  The second
instance fails to acquire the lock that is still held by the first
instance so it exits.  Eventually, the first instance also exits and
we're left with no dhcpcd running.  This CL waits for dhcpcd to exit
after we send it a SIGTERM.

BUG=chromium-os:26580
TEST=Unit tests, run network_WiFiManager and examine the messages to
ensure that processing does not continue until after dhcpcd exits

Change-Id: I8b4c9ea21ec652eb36f69b37841424b72c0f6078
Reviewed-on: https://gerrit.chromium.org/gerrit/16323
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Ready: Thieu Le <thieule@chromium.org>
Tested-by: Thieu Le <thieule@chromium.org>