gobi_modem: recognize registered modems at cromo start
When cromo starts it checks the power state of the modem to see if the
modem is already enabled.  If it is already enabled it sets the
internal state to indicate this.   It also needs to check to see if
the modem is already registered on the cellular network, and if so,
put the modem into the "registered" state.

BUG=chromium-os:26791
TEST=start and start cromo without flimfliam running.  Observe modem
     State with the modem status command

Change-Id: I590325524eb6ddbedf0a2ee6f51358ff4af04e78