[cryptohome] [rfc] switch to Make
building cryptohome right now takes 35 seconds, and building and running tests
takes nearly 5 minutes. We rebuild each object file 3-4 times as part of
testing.

Also, remove SecureBlob's DestructorTest, which fails reliably when cryptohome
is built using make. The test uses memory after free, so diagnosing failures of
it is an exercise in frustration; I'm removing it and declaring it dead.

Change-Id: Icc32a4b559466d1bd113c4b4b9024cca547276a0
Signed-off-by: Elly Jones <ellyjones@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/16519