shill: separate (externally observed) failure state from internal failure condition.
Changing a service's state to failure causes Chrome to display
a connection failed dialog. Consequently, before this CL, if
a user suspends a device at home, and resumes it at the office,
he'll see a spurious connection failure dialog. (shill will
display the dialog when it can't connect to the home AP, even
though it is about to connect to the office AP.)

To resolve this, we set a WiFiService's state to idle, rather
than failure. However, internally, we record the time at which
the service disconnected, and use that to determine whether or
not a service is failed.

BUG=chromium-os:26828
TEST=new unit tests

Change-Id: I556a110f5705495d681f40c1bb645c34bd0e43f9
Reviewed-on: https://gerrit.chromium.org/gerrit/18685
Commit-Ready: mukesh agrawal <quiche@chromium.org>
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>