debugd: add systrace support
Add simple support for doing system tracing.  SystraceStart registers a
set of kernel events and enables collection using the global trace clock.
SystraceStop ends collection, clears the registered events, and writes
the collected data to stdout (which the caller should attach to a file or
similar). SystraceStatus reports the current state of event collection
in an unstructued (text) format (suitable for sending directly to the
terminal).

Note the helper assumes debugd has write-access to files in the tracing
debugfs directory.

BUG=chromium-os:27809
TEST=use crosh systrace support to invoke start, status, and stop methods

Change-Id: I42e1f3c6ba3b543e608a358bb4682b1a7f8ba02f