Temporarily disable dutflaga/b GPIO export on all platforms.
We keep seeing a problem with unexpected signals read on these exported
GPIOs on several boards (just saw it on Lumpy, too).  Currently, Alex is
the only board we actually trust in this respect, but for the time being
we'll still disable dutflaga/b GPIO export across the board.  We should
also implement a safe fallback mechanism for the server URL prior to
re-enabling GPIO export for individual boards.

BUG=chromium-os:27077
TEST=Solves the problem seen on ZGB/Lumpy.

Change-Id: Id31ba390eae59a4d3581eb3725313e51807e018b
Reviewed-on: https://gerrit.chromium.org/gerrit/17165
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Tested-by: Gilad Arnold <garnold@chromium.org>
Commit-Ready: Gilad Arnold <garnold@chromium.org>