Fixing a subtle bug in powerd's protobuf over dbus code.
The value that is supposed to be passed into dbus for an array is not the array
pointer, but actually a pointer to the array pointer, o_O. In this case this was
causing a segfault in powerd whenever a call to the new protobuf based path was
occuring. I did not catch this originally, since I had both paths running in my
test, so the data was still making its way to chrome and powerd was being
silently restarted. Using just the new protobuf path is when issues were seen.

BUG=chromium-os:25489
TEST=Built image using chrome that had this CL and
     http://codereview.chromium.org/9318007/ in it.
     Observed that data still arrived for the power status icon and powerd was
     not crashing in dmesg

Change-Id: I20ec6a8c0219ffc0955371e9fc460603f8dbc810
Reviewed-on: https://gerrit.chromium.org/gerrit/17175
Reviewed-by: Satoru Takabayashi <satorux@chromium.org>
Commit-Ready: Ryan Harrison <rharrison@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>