[cryptohome] allow MountTaskObservers to clean themselves up.
Allow the MountTaskObserve method on MountTaskObserver to return true to
indicate that the caller (usually a MountTask) should free the
MountTaskObserver. This makes it possible to create new MountTaskObserver
subclasses which encapsulate state we need to have hang around without problems
cleaning them up. Without this change, the call flow goes:

MountTask
-> MountTaskObserve
   -> ...observer code...

The observer code can't free the Observer because we're about to return through
it.

BUG=chromium-os:20778
TEST=unit

Change-Id: Ife0ba88fe4bdc255fcfb69ddc8c29d538d02e44e
Signed-off-by: Elly Jones <ellyjones@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/17359
Reviewed-by: Kees Cook <keescook@chromium.org>