Make session state change code aware of the previous session state
In the logs for powerd I was noticing that there was multiple messages during
the start up sequence telling me about metrics that couldn't be generated and
spurious metrics being sent. Investigating I determined the root cause of this
is that during the initialization method we are retrieving the session state,
which returning stopped. We then proceed to run the change of state method,
which causes the metrics generation method to be called. Since a start has not
been processed the metric counter values are not set, leading to the weird
behaviour.

This CL adds a variable for the current state and checks the value when we call
the change state method. If we are already in the state that we are going into
we skip the actions and treat it like a no-op, unless it is the started. For
started we record an error and proceed as normal. I also added in a logging
message, so that we know what we retrieved, so all of the session state
information that we have received appears in the log.

BUG=chromium-os:27470
BUG=chromium-os:27471
TEST=Built image using CL. Looked in the file
     /var/log/power_management/powerd.LATEST and saw the messages that are
     mentioned in the bug are not longer present. Confirmed the retrieved
     message appears as desired and there are not new error messages.

Change-Id: Ie9e83ff96970de8db212be81428218f69920fd11
Reviewed-on: https://gerrit.chromium.org/gerrit/17502
Tested-by: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Simon Que <sque@chromium.org>
Commit-Ready: Ryan Harrison <rharrison@chromium.org>