[cryptohome] return JSON from GetStatusString.
Add support for emitting the state of multiple mounts in GetStatusString. As
part of this, move the code that emits the status of a particular mount out into
Mount where it belongs. Do the same with the TPM and InstallAttributes status
code. To avoid returning strings of an adhoc format that we paste together,
instead return JSON values and return a pretty-printed JSON string.

TODO: Unit tests

BUG=chromium-os:20778
TEST=trybot

Change-Id: Idf4d00a6f9e01fa51326e69b9f0a8f5c44b6d6fc
Signed-off-by: Elly Jones <ellyjones@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/17516
Reviewed-by: Kees Cook <keescook@chromium.org>