Convert RollingAverage to int64
The commited version of the RollingAverage class uses doubles as the interior
type. This was a compromise since I was intending on using it for both floating
point and integer data. I have decided to not use it for the floating point
data, so I am converting it to the correct integer type. If there is a need in
the future for it to support multiple data types I would consider making it a
template class, but I think that is too heavy weight of a solution at the
moment.

BUG=chromium-os:23647
TEST=Confirmed that the code still builds and the unit tests pass.

Change-Id: I1cf924eaa3b0be17c9f16ac813f2b871d4d581b1
Reviewed-on: https://gerrit.chromium.org/gerrit/17524
Tested-by: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Simon Que <sque@chromium.org>
Commit-Ready: Ryan Harrison <rharrison@chromium.org>