Created a simple class to help clients send login events.
The ChapsProxyImpl class is large and pulls in a bunch of stuff internal
to Chaps. It's not well-suited to be used by other projects. The
LoginEventClient class, by contrast, requires only a single header to be
published and provides a simple interface for firing off events to the
Chaps daemon.

Also changed the D-Bus interface to pass authorization data as a byte
array rather than a string.

BUG=chromium-os:21003
TEST=Manually tested with a client that sends login events.

Change-Id: Iba34d7092c7741c8b66d5bea384cbebff0ee3855