Fixing black external monitor bugs
Currently we have a bunch of black external monitor bugs: Alex with VGA/VGA
dongle output, Lumpy with DP to DP/VGA/HDMI output. When system is booted
with external monitors attached, sometimes the monitor is black. And also
the policy that LCD should be off when external monitor has higher resolution
is not followed on Alex.

I made some changes to monitor_reconfigure to address the issues above, and
the main changes are:

1. Disable all the outputs, both connected and disconnected, before reconfigure,
instead of just disconnected outputs. This makes the logic simple: disable all
outputs, and then enable the outputs we need.

2. In addition to listening to RRScreenNotifyEvent, also listen to
RROutputChangeNotifyEvent. This makes the detection of monitor hotplug more robust.

3. Don't run reconfigure on duplicate RR*NotifyEvent. Since RR*NotifyEvent
can be generated during the monitor reconfigure and causes monitor_reconfigure
to Run() again, results in duplicate calls or possible endless loop. Adding
the condition that if RR*NotifyEvent is received within a small abount of time
since last reconfiguration and the number of outputs don't change, then ignore
the events.

4. Adding logic to not reconfigure to handle special cases like first time laptop
boots without external monitor, or wake up from suspend, so to avoid unpleasant
blinking.

5. Enforcing the policy of disalbing LCD output if external monitor has higher
resolution.

6. Removing DPMS call from backlight_controller since directly operating on DPMS
is risky on ARM platform (as marcheu@ pointed out). Only keeping one DPMS call in
monitor_reconfigure after enabling usable outputs. This is a workaround of a
XRandr/DRM driver bug that sometimes using XRRSetCrtcConfig() to disable outputs
turns off the CRTC and DPMS but later on using XRRSetCrtcConfig() to enable the
same output the CRTC and DPMS are not correctly restored. So we always call DPMS
to turn back on the CRTC and DPMS after enabling usable outputs.

7. To replace DPMS, use MonitorReconfigure::SetScreenOff() to turn off screen when
system goes into IDLE_OFF state, and use MonitorReconfigure::SetScreenOn to turn on
the screen when coming back from IDLE_OFF and SUSPENDED states.

BUG=chrome-os-partner:8157, chromium-os:17743, chrome-os-partner:7088
TEST=Run the following tests on R19 with Alex/Lumpy/Stumpy
     1. Connect device with external monitor,
        VGA/VGA dongle for Alex,
        DP -> DP/VGA/HDMI for Lumpy
        DP -> DP for Stumpy
        Boot device, and make sure external monitor comes up, and
        LCD is off (for laptop).
        Disconnect external monitor, make sure screen comes back to LCD.
        (for laptop)
     2. Boot device first, make sure screen comes up to LCD.
        Connect external monitor, and make sure LCD is off and external
        monitor shows chromeos.
        Disconnect external monitor, make sure screen comes back to LCD.
     3. Wait until the device goes into IDLE_OFF state and make sure the
        the screen is turned off. When waking up from IDLE_OFF or suspend
        state, make sure the screen comes back.

Change-Id: Icedad755bd4d3f9a6e33cea85ff453a285ce8b50
Reviewed-on: https://gerrit.chromium.org/gerrit/17877
Commit-Ready: Yufeng Shen <miletus@chromium.org>
Reviewed-by: Yufeng Shen <miletus@chromium.org>
Tested-by: Yufeng Shen <miletus@chromium.org>