Search for all the possible resume strings
This CL refactors how we search for time stamps related to kernel susspend and
resume time. It allows for easier support of multiple potential message strings
and makes the sections more independent, so that if one metric doesn't work we
can still get others. It also adds in error messages related to failing to
generate values. I took inspiration from the behaviour that occurs in the
power_Resume test, since it handles the changing kernel string problem already.

This should resolve issues with generating the various KernelResume/SuspendTime*
metrics, but not the Firmware* ones.

BUG=chromium-os:13590
TEST=Built image with CL and followed the reproduction instructions from the bug
     with one exception. I waited 30+ seconds, since the refresh can be that
     long. The histograms generated as expected for the KernelResume/Suspend*.

Change-Id: I2b23a84bc6400c6eb57dd492af3ce1f110c05fbe
Reviewed-on: https://gerrit.chromium.org/gerrit/18471
Tested-by: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Darin Petkov <petkov@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Commit-Ready: Ryan Harrison <rharrison@chromium.org>