[cryptohome] move credential tests off Mount
Move the vast majority of TestCredentials off Mount and onto HomeDirs. After
this change, Mount's TestCredentials is a stub which checks the supplied
credentials against the held ones (if there are held ones) and does nothing
else, and HomeDir's AreCredentialsValid() actually tests against the data on
disk.

Some of the unit test changes this causes are *extremely* gruesome. Some of the
unit tests want to check that a home directory they have created (using Mount)
has valid credentials in it, which necessitates a call to HomeDirs, so they now
have to set up a HomeDirs configured roughly equivalently to their Mount. It is
my fervent hope that this will be refactored away some day.

BUG=chromium-os:20778
TEST=unit,trybot

Change-Id: I366e8379c9e354329458efc1d6aade7753bdd78c
Signed-off-by: Elly Jones <ellyjones@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/18266