Update dbus policy and run as 'chaps'.
The dbus interface can now be owned by 'chaps' or 'root'. All members of
the 'pkcs11' group have normal access to the interface. When chaps is
disabled and calls are redirected to opencryptoki, the daemon still runs
as root and effective chronos. When chaps is enabled, the daemon will
switch to real chaps on startup. Cryptohome will be responsible for
creating a token database directory which is accessible by the chaps
user.

BUG=chromium-os:27967
TEST=Run all unit tests.
     Manually test with chaps enabled and disabled:
     - Use PKCS #11 services as chronos.
     - Use PKCS #11 services as wpa.
     - Connect to Google-A.

Change-Id: I24f29e6b7dab91c07bc5f428598c03653e35a748