shill: Implement Service::Configure
This applies the properties of a KeyValueStore to a Service.  It
does so by calling accessors in the Service's PropertyStore with
each value in the KeyValueStore.  Certain parameters are ommitted
at the request of the Service if they were ostensibly already handled
in the GetService or ConfigureService call that started this Confiugre.

BUG=chromium-os:22800
TEST=New unit tests.

Change-Id: I69eedcdee03fa0682ff7821553bb6c7d76e6a50e
Reviewed-on: https://gerrit.chromium.org/gerrit/18823
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Ready: Paul Stewart <pstew@chromium.org>