Removing tracked directory migration
I has been a year since the migration was added. It was used for ~/Downloads only.

A user may need the migration if
(i) she has valuable data in her ~/Downloads,
(ii) it was created prior to the migration added (a year ago) and
(iii) she has not undertook the migration since then.

BUG=chromium-os:14686
TEST=FEATURES="test" emerge-x86-alex chromeos-cryptohome

Change-Id: I350b4ece83eede354cd08fa75c4e4a99e7b13554
Reviewed-on: https://gerrit.chromium.org/gerrit/18824
Tested-by: Denis Glotov <glotov@chromium.org>
Reviewed-by: Nikita Kostylev <nkostylev@chromium.org>
Commit-Ready: Denis Glotov <glotov@chromium.org>