powerman: Turn off error logging on unhandled messages
Before the refactoring (gerrit #16322), there was no logging when the
handler got a message without a case for handling it.  Now there is such
logging, and it pollutes the logs needlessly.  This patch cleans it up.

In the future, we should further refactor the dbus handling so that the
handler doesn't get called at all for these unhandled messages.

BUG=chromium-os:26674
TEST=power_manager builds successfully

Change-Id: Ifb5c7fb0115cfc892ab03a70ac538a892b2399ce
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/18907
Reviewed-by: Daniel Erat <derat@chromium.org>